Move active taker to one db
